refactor: extract shared StatusList mapper to eliminate 3x duplication (#79)
Create toStatusListDomain() in lib/actions/mappers.ts and replace
three identical StatusListRow→StatusListDomain mappings in board.ts
(getStatusLists, createDefaultStatusLists, createStatusList).
